Completed Project 1
For this first project I completed the two different pages. Inside the index.html, I placed some reviews from customers and also linked an image to go along with the reviews. A navigation menu was also created for the user to switch between the different sections of the page. Some basic information such as the email and phone number was added. For the jtc_services.html, I added the services that the company provides as well as some basic information about them. At the bottom the address of the company was also added inside a footer element